scraper/pyproject.toml
[project]
name = "au-recycling-scraper"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32",
    "beautifulsoup4>=4.12",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
    "lxml>=5.3",
    "mcp>=1.0",
    "pyyaml>=6.0.3",
    "anthropic>=0.84.0",
    "playwright>=1.58.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
